### 2020-06-12 #2333 @whedon
[PRE REVIEW]: rsudp: A Python package for real-time seismic monitoring with Raspberry Shake instruments
**Submitting author:** @iannesbitt (<a href="http://orcid.org/0000-0001-5828-6070">Ian Nesbitt</a>)
**Repository:** <a href="https://github.com/raspishake/rsudp" target ="_blank">https://github.com/raspishake/rsudp</a>
**Version:** v1.0.1
