### 2020-06-18 #2333 @Kevin-Mattheus-Moerman
@kbarnhart would you be okay with me assigning this paper to you now, but that it is okay that you pick this up once you have time e.g. in 1-2 weeks? It is acceptable that we are slower handling things given our "reduced service mode", so we can wait until you are ready.
