### 2020-08-02 #2333 @arfon
@jedbrown ☝️ looks like https://github.com/openjournals/whedon-api/pull/108 works nicely.
