### 2020-08-14 #2565 @whedon
[REVIEW]: rsudp: A Python package for real-time seismic monitoring with Raspberry Shake instruments
**Submitting author:** @iannesbitt (<a href="http://orcid.org/0000-0001-5828-6070">Ian Nesbitt</a>)
**Repository:** <a href="https://github.com/raspishake/rsudp" target ="_blank">https://github.com/raspishake/rsudp</a>
**Version:** 1.1.0

@calum-chamberlain and @fwalter, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
