### 2020-09-11 f5ede27
more specific errors related to invalid miniseed files
