### 2020-11-16 #2565 @iannesbitt
note to reviewers, here is a potentially relevant recent citation that uses RS that might be good to include among the first few references cited:

https://science.sciencemag.org/content/369/6509/1338.abstract
