### 2020-11-18 #2565 @jedbrown
@fwalter (Fabian Walter) :wave: Welcome to JOSS and thanks for agreeing to review!

The comments from @whedon above outline the review process, which takes place in this thread (possibly with issues filed in the [rsudp repository](https://github.com/raspishake/rsudp/)). I'll be watching this thread if you have any questions.

We aim for reviews to be completed within a month or so. Please let me know if you require some more time. We can also use Whedon (our bot) to set automatic reminders if you know you'll be away for a known period of time.
