### 2020-12-24 #2565 @fwalter
Dear Jed,

I have reviewed this software package and would now like to submit my
report. Do I understand correctly that this should be done as an issue on
the repository:

https://github.com/raspishake/rsudp/issues

Also, where can I fill out the checklist, which is mentioned here:

https://joss.readthedocs.io/en/latest/review_checklist.html

I tried to accept the invitation by whedon, but this seems to have expired.
Is that a problem?

In general, this software package compiles and functions well, but it seems
to be mostly (perhaps only) useful for educational purposes. For research
or monitoring purposes one has to include custom modules, which I found
tedious, especially when the user is not highly fluent in python.

Thanks for your help!
Fabian.

Am Mi., 18. Nov. 2020 um 02:17 Uhr schrieb Jed Brown <
notifications@github.com>:

> @fwalter <https://github.com/fwalter> (Fabian Walter) 👋 Welcome to JOSS
> and thanks for agreeing to review!
>
> The comments from @whedon <https://github.com/whedon> above outline the
> review process, which takes place in this thread (possibly with issues
> filed in the rsudp repository <https://github.com/raspishake/rsudp/>).
> I'll be watching this thread if you have any questions.
>
> The JOSS review is different from most other journals. Our goal is to work
> with the authors to help them meet our criteria instead of merely passing
> judgment on the submission. As such, the reviewers are encouraged to submit
> issues and pull requests on the software repository. When doing so, please
> mention #2565 <https://github.com/openjournals/joss-reviews/issues/2565>
> so that a link is created to this thread (and I can keep an eye on what is
> happening). Please also feel free to comment and ask questions on this
> thread. In my experience, it is better to post
> comments/questions/suggestions as you come across them instead of waiting
> until you've reviewed the entire package.
>
> We aim for reviews to be completed within a month or so. Please let me
> know if you require some more time. We can also use Whedon (our bot) to set
> automatic reminders if you know you'll be away for a known period of time.
>
> Please feel free to ping me (@jedbrown <https://github.com/jedbrown>) if
> you have any questions/concerns.
>
