### 2020-12-24 #2565 @jedbrown
@fwalter Thanks! When you accept that fresh invitation, you should be able to click in the checkboxes.

Indeed, general comments can go directly in this thread, but specific "to do" items should be filed as issues in https://github.com/raspishake/rsudp/issues. They'll be automatically linked from here if you mention this issue (`https://github.com/openjournals/joss-reviews/issues/2565`).
