### 2020-12-30 raspishake/rsudp#12 @fwalter
Improve data handling
Hi there,

It would help to have a function, which converts data from the queue to an obspy stream object, which can then be manipulated by the user. This would immediately allow the user to access the entire obspy functionality, which is arguably the most widely used and powerful open source package for seismic data handling. Such a function should also bypass the module manipulation required to write custom consumers as currently outlined in the documentation.

[https://github.com/openjournals/joss-reviews/issues/2565](url)
