### 2020-12-30 raspishake/rsudp#13 @fwalter
Some minor remarks on the documentation
Hi again,

MINOR TYPOS IN DOCUMENTATION AND CODE MESSAGES:

"You canthen  enter the rsudp conda environment by typing "conda activate rsudp" and then run rsudp by using the command "rs-client -h""

"The client then instantiates the relevant threading.Thread objects as and passes"

In the configuration file and documentation, the sound setting is sometimes referred to as "alarm sound" and sometimes as "alert sound". I assume these are the same?

REMARKS ON DOCUMENTATION

6. Troubleshooting rsudp: "Once you have a good setup worked out, you will typically be able to start and run this program without errors.": This is a somewhat trivial statement and may annoy users who are not getting their Raspberry devices running. I suggest being more specific here or removing this sentence.

1.2.1. Message types: why not document "data"? As a non-python expert (though frequent user) I spent some time googling what the data types starting with a "b" mean and how they have to be handled. Then I had to find the correct methods in the rsudp modules. It would be easier to provide a function that does all this for users only interested in the data stream (via an obspy stream object, e.g., see my other issue [https://github.com/raspishake/rsudp/issues/12](url) ).

[https://github.com/openjournals/joss-reviews/issues/2565](url)
