### 2020-12-30 raspishake/rsudp#13 @fwalter
One more comment: When changing the python modules I noticed that the changed code is not executed unless the start-up script is re-run. Again, I am not a python expert, so this may be obvious, but it would have helped me to see this in the documentation.
