### 2020-12-30 raspishake/rsudp#13 @iannesbitt
These are very helpful fixes, thank you. I will work on incorporating them.
