### 2020-12-31 raspishake/rsudp#12 @fwalter
Hello @iannesbitt, I was thinking about something like this:

https://docs.obspy.org/tutorial/code_snippets/easyseedlink.html

Otherwise, would it be possible to return some other type of python variable or object, which updates itself every time new data are available?
