### 2020-12-31 #2565 @fwalter
The submitted python-based rsupd package is designed to handle data output of the Raspberry Shake and Raspberry Boom sensors. Although I have not used these sensors before I understand that they are user friendly and low cost seismic and infrasound sensors particularly suitable for educational purposes or research applications, where specialized custom-made instrumentation is necessary.

The code can be installed and compiled with no problems and functionalities and python classes are documented well. As I do not own a Raspberry Shake or Raspberry Boom myself, I had to rely on the testing and demoing mode for this review. I have no reason to doubt that the package works well with the actual hardware, but I would suggest including a step-by-step tutorial on how to set up, handle and configure the software together with the hardware. The documentation does include a section on how to run the software in batch processing mode.

Beyond data reading from a Raspberry instrument, the software package does not seem to offer new scientific functionality compared to existing software. Spectrograms, amplitude statistics and STA/LTA detectors are standard in high-level seismic analysis toolboxes (e.g. obspy, GISMO, SEIZMO) or can be easily programmed from scratch. The twitter and telegram functionality can be a useful add on for communicating realtime processing output.

As outlined in the “Developer’s Guide”, the program architecture seems straightforward and appropriately designed. The definition and purpose of queues and threads and the difference behind these concepts may not be clear to all users. I suggest elaborating on this to allow a non-software-expert grasp the operation of rsupd better.

From the paper it is not clear how this software compares with other packages (hence the unchecked box in the Review List). There is a paragraph on this in the documentation (“1.2. Why It’s Special”), however, the argumentation is very general. The code requirement that “calculations be made quickly” should be elaborated on in this regard (Which kind of calculations? How/to what extent can they already be done in other packages?).

From my point of view, this is a particularly helpful software package for teaching purposes. I can envision student exercises using this software package in combination with the Raspberry hardware. Beyond the classroom this could be useful for applications where Raspberry products run operationally and continuously. Personally, I would not switch to this approach, as for my research I stream and archive seismic data via professional national network agencies. For Raspberry users, the rsupd package may help data handling, but I found the described solution to include custom consumers somewhat difficult to use as several rsupd modules required modification (although the given instructions were clear and worked well). As described in issue [#12](https://github.com/raspishake/rsudp/issues/12) on the rsudp github page, it would be helpful to have a single function, which packages data into a python object, ideally an obspy stream object. Users should be able to hook into such an interface without detailed knowledge of the source code or expert programming knowledge. This would give access to the most advanced open source and tested code for analyzing seismic data and the obspy package is already used for making rsupd run, so this would be a natural choice.

In summary, this is a useful software package for educational purposes and customers of the Raspberry products. I expect Raspberry Shake/Boom users to employ and extend this package in their applications. If the data handling interface is simplified and improved (see issue [#12](https://github.com/raspishake/rsudp/issues/12) on the target repository), this software package will facilitate realtime data handling for Raspberry users.

Note: The references do not appear in the software paper.
