### 2020-12-31 #2565 @jedbrown
Thanks, @fwalter. What do you mean by "The references do not appear in the software paper."? I see 8 references with DOI and nothing unresolved in the PDF.

I suppose I would mention that even if something can be easily coded from scratch, there's still value in having an authoritative implementation for provenance, accessibility, and sometimes performance. The ObsPy integration you suggest sounds like it would be useful.

@iannesbitt Please let us know if you have questions about revising to address the concerns above.
