### 2020-12-31 #2565 @jedbrown
It was linked above in the thread, but here's a fresh version. ☝🏼
