### 2021-01-05 cc1b397
updating docs for #14
