### 2021-01-07 6d6ba22
correcting link destinations in social media modules
