### 2021-01-07 19ac6ba
adding changes for #14 and #15 and a version bump
