+++ CHANGES.md
@@ -1,4 +1,8 @@
 # Changelog
+## changes in 1.0.2
+- corrected install script to fix [#14](https://github.com/raspishake/rsudp/issues/14)
+- corrected social media URL destination [#15](https://github.com/raspishake/rsudp/issues/15)
+
 ## changes in 1.0.1
 - added `rsudp.c_rsam` Real-time Seismic Amplitude Measurement consumer
 - modified install scripts for clarity
