### 2021-01-08 825c683
implementing #9, extra static text for twitter messages
