+++ docsrc/settings.rst
@@ -407,7 +407,33 @@ If :json:`"enabled"` is :json:`true`, and all API keys have been generated and a
 then the :class:`rsudp.c_tweet.Tweeter` class will use the Twitter API to
 create tweets when an ALARM message arrives on the queue.
 If :json:`"tweet_images"` is :json:`true`, then the module will also tweet a saved image of the event,
-if :json:`"eq_screenshots"` is set to :json:`true` in the "plot" module.
+if :json:`"eq_screenshots"` is set to :json:`true` in the "plot" module. If any text is put in the
+:json:`"extra_text"` string, then the software will insert that text (no longer than 143 characters)
+into the tweets after a single space. For example, an unmodified tweet with :code:`"extra_text": ""`
+might look like this:
+
+.. _eq-tweet:
+.. figure::  _static/eq_tweet.png
+    :align:   center
+
+    An example tweet sent with the "extra_text" parameter empty (this is the default).
+
+
+Changing the :json:`"extra_text"` parameter to :code:`"extra_text": "from #Williamstown #MA"`
+would render something like the following:
+
+.. _eq-tweet-extra:
+.. figure::  _static/eq_tweet_extra.png
+    :align:   center
+
+    An example tweet sent with the "extra_text" parameter filled.
+
+.. warning::
+
+    Starting the software with an :json:`"extra_text"` string in excess of 143 characters
+    will yield a warning and the :json:`"extra_text"` string will be truncated
+    in order to avoid the tweet being rejected for exceeding the 280 character limit.
+
 
 .. _setting-up-twitter:
 
