### 2021-01-08 c9162e9
fixing bug with implementation of #9
