### 2021-01-08 07ebd5a
fixing bug with implementation of #9
