### 2021-01-19 fe9fcff
correcting extra_text based on longest flinn-engdahl names
