### 2021-01-19 a00947b
documenting addition of feature requested in #9
