### 2021-01-19 8983ce7
adding documentation changes requested in #13
