+++ README.md
@@ -15,10 +15,10 @@
 `rsudp` has [full documentation here](https://raspishake.github.io/rsudp/). We also have [tutorial instructions](https://raspishake.github.io/rsudp/index.html#tutorial) to install, set up, and run rsudp there. Additionally, our documentation features [YouTube walkthroughs](https://raspishake.github.io/rsudp/youtube.html), a brief [Developer's guide](https://raspishake.github.io/rsudp/theory.html), and [module documentation](https://raspishake.github.io/rsudp/#code-documentation).
 
 `rsudp` contains ten main features:
-1. **Alarm** - an earthquake/sudden motion alert trigger, complete with a bandpass filter and stream deconvolution capabilities
-2. **AlertSound** - a thread that plays a MP3 audio file in the event of an alarm
+1. **Alert** - an earthquake/sudden motion alert trigger, complete with a bandpass filter and stream deconvolution capabilities
+2. **AlertSound** - a thread that plays a MP3 audio file in the event of the alert module signalling an alarm state
 3. **Plot** - a live-plotting routine to display data as it arrives on the port, with an option to save plots some time after an alarm
-4. **Tweeter** - a thread that broadcasts a Twitter message when the alarm module is triggered, and optionally can tweet saved plots from the plot module
+4. **Tweeter** - a thread that broadcasts a Twitter message when the alert module is triggered, and optionally can tweet saved plots from the plot module
 5. **Telegrammer** - a thread similar to the Tweeter module that sends a [Telegram](https://telegram.org) message when an alarm is triggered, which can also broadcast saved images
 6. **Writer** - a simple miniSEED writer
 7. **Forward** - forward a data cast to another IP/port destination
