+++ CHANGES.md
@@ -3,6 +3,7 @@
 - corrected install script to fix [#14](https://github.com/raspishake/rsudp/issues/14)
 - corrected social media URL destination [#15](https://github.com/raspishake/rsudp/issues/15)
 - adding feature requested in [#9](https://github.com/raspishake/rsudp/issues/9) (additional text for twitter posts [documented here](https://raspishake.github.io/rsudp/settings.html#tweets-twitter-notification-module))
+- edited language as requested in [#13](https://github.com/raspishake/rsudp/issues/13)
 
 ## changes in 1.0.1
 - added `rsudp.c_rsam` Real-time Seismic Amplitude Measurement consumer
