### 2021-01-19 #2565 @iannesbitt
I have implemented the changes requested by @fwalter in raspishake/rsudp#13.
