### 2021-01-19 raspishake/rsudp#13 @iannesbitt
These suggestions have been implemented:

- https://github.com/raspishake/rsudp/blob/8983ce7d74aad3d1508e42119b25d9bdc5dcd62f/bin/unix-install.sh#L269
- https://raspishake.github.io/rsudp/theory.html
- https://github.com/raspishake/rsudp#rsudp
- https://raspishake.github.io/rsudp/troubleshooting.html
- https://raspishake.github.io/rsudp/settings.html
