### 2021-01-19 17cc70b
adding extra_text to default settings documentation
