+++ docs/_sources/settings.rst.txt
@@ -619,7 +619,8 @@ By default, the settings are as follows:
         "api_key": "n/a",
         "api_secret": "n/a",
         "access_token": "n/a",
-        "access_secret": "n/a"},
+        "access_secret": "n/a",
+	    "extra_text": ""},
     "telegram": {
         "enabled": false,
         "send_images": true,
