### 2021-01-19 412f03c
adding extra_text to default settings documentation
