### 2021-01-20 29b8f19
adding versionadded for extra_text setting
