### 2021-01-20 3b6ffe4
adding feature requested in #17
