+++ docs/_sources/settings.rst.txt
@@ -496,6 +496,9 @@ accordingly. Setting :code:`"fwd_data": true` will forward data from the specifi
 :code:`"fwd_alarms": true` will forward :code:`ALARM` and :code:`RESET` messages. These can both be set to true
 simultaneously.
 
+To take advantage of this forwarding capability in another piece of software, it may help to consult the
+:ref:`message-types`.
+
 `Back to top ↑ <#top>`_
 
 
