### 2021-01-20 9881017
adding ability to forward to a list of destinations (#18)
