### 2021-01-20 fd53647
adding #18 to change log
