### 2021-01-20 0034a2a
adding #18 to readme
