### 2021-01-22 3365188
fixing bug in miniseed writer
