### 2021-01-23 79aa8f4
adding stream builder consumer
