### 2021-04-19 8f3c1eb
making log creation more downstream-friendly
