### 2021-04-19 66b7ab1
making log creation more downstream-friendly
