### 2021-08-11 raspishake/rsudp#12 @calum-chamberlain
If it helps, I had to cope with handling streaming data (that may also be gappy) and converting to ObsPy objects in RT-EQcorrscan - the solution I have is [here](https://github.com/eqcorrscan/RT_EQcorrscan/blob/master/rt_eqcorrscan/streaming/buffers.py), and doesn't seem to be leaky. There is then a painful array of `Queues` to handle getting the current state of the `Stream` from the streamer.
