### 2021-08-11 #2565 @calum-chamberlain
This package provides a useful, simple to use and attractive real-time interface to RaspberryShake data, and I look forward to using it as an educational and outreach tool within my University and in broader communities. Firstly I would like to thank the developers for this tool. I do own a raspberry shake with a single-component sensor and have tested this with my own shake. Setup was simple and the instructions were easy to follow. I would have liked to see a link to setting up the UDP server on the shake from within the [installation](https://raspishake.github.io/rsudp/installing.html) docs.

While the code seems to be of good quality, and has a nice simple test for functionality, it would be good to see unit-tests of some of the modular elements that could be run on CI systems to ensure that future changes to the code do not break any functionality. This may also make it easier for others to contribute extensions to this code, who may not be so familiar with all the elements of the package. On a similar note, I did not see any contributing guidelines on the github page, and would like to see a link to the [Contributing to this project](https://raspishake.github.io/rsudp/contributing.html) page from the github repository.

To me this code would be very helpful for education and outreach, and the statement of need in the online documentation is good, however I think the paper would benefit from taking some of the text from the [Why it is special](https://raspishake.github.io/rsudp/about.html#why-it-s-special) section of the docs to make the justification for the code more obvious. It would also be helpful to compare to other software for a similar purpose – I note that the “About” page of the documentation states that “Programs that do similar tasks are usually not as fully-featured, cost money, are unmaintained, or are complex to set up and run.” however it would be good to reference what those software are and have this short comparison statement in the paper.

As eluded to by @fwalter – being about to access data in an ObsPy Stream would be useful, although for most users I imagine that the standard configuration as offered now will be sufficient for their needs. However, for shakes connected to the internet data are available via [FDSN](https://www.fdsn.org/networks/detail/AM/), and this would be the natural route for collecting data for most RaspberryShake installs. It may however be useful to have access to an ObsPy Stream for storing data on a separate co-located machine in non-standard, not internet-enabled installs. I do not know how common such installs are though.

In summary, this is a nice visualisation tool for a specific set of seismic instruments that will likely be well-used by raspberry shake owners for outreach and education.
