### 2021-08-11 #2565 @jedbrown
Thanks for your review, @calum-chamberlain. Could you clarify if/how you consider this useful for _research_, not strictly outreach and  education? Are there parts of functionality or documentation that should be improved to better meet the needs of researchers?
