### 2021-08-11 #2565 @calum-chamberlain
Hi @jedbrown - the research need of this software is not particularly clear to me given that data can be accessed from via seedlink and accessible via other tools, and hence through ObsPy as is commonly used by "power users". The ability to archive miniseed data locally is helpful, but as these continuous data are available to researchers via FDSN, alongside other data that would likely be useful for research, it isn't clear to me why I would want to archive data locally, unless (as suggested above) the raspberry shake installation was deployed without an internet connection and data were to be retrieved intermittently.

That this software provides a means to access the UDP Datacast is helpful, but it would be helpful to know if/why this is preferred to the seedlink stream.

A caveat to this is that my research needs, and therefore my view of this, might be narrow and missing something and I would be happy to be corrected!
