### 2021-08-11 38a95c4
adding link to contributor notes
