+++ paper/paper.bib
@@ -1,6 +1,7 @@
 @article{Anthony:2018,
     author 		= {Anthony, Robert E. and Ringler, Adam T. and Wilson, David C. and Wolin, Emily},
-    title 		= "{Do Low‐Cost Seismographs Perform Well Enough for Your Network? An Overview of Laboratory Tests and Field Observations of the OSOP Raspberry Shake 4D}",
+    title 		= "{Do Low‐Cost Seismographs Perform Well Enough for Your Network? An Overview of Laboratory Tests
+					and Field Observations of the OSOP Raspberry Shake 4D}",
     journal 	= {Seismological Research Letters},
     volume 		= {90},
     number 		= {1},
@@ -40,7 +41,7 @@
 	URL			= {https://doi.org/10.3389/feart.2020.00009},
 	DOI			= {10.3389/feart.2020.00009},
 	ISSN		= {2296-6463},
-	ABSTRACT={The high degree of human activities in urban environments produces large background vibrations that makes it difficult to use data acquired in these areas for classical seismology. Seismometers installed within cities have been typically been used for the study of seismic hazard or for monitoring civil engineering problems. However, with the development of monitoring techniques based on the interpretation of the so-called seismic ambient noise, these data have gained scientific interest. Our objective is to discuss an additional utility of seismometers deployed within a city; its use as a tool to connect society with Earth sciences. Many citizen activities, from traffic to music concerts, produce vibrations that can be recorded seismically, and our experience shows that these records attract the attention of the media and social networks. With the emergence of low-cost and easy-to-use instruments in recent years, more citizens can now record ground motion and become interested in the interpretation of the recorded seismograms. The installation of permanent seismic networks in educational centers has proven to be a good approach to introduce students to Earth sciences at the national level and can also be developed at the urban scale using this new instrumentation. In this contribution we will first review the previous results related to the identification of the sources of vibration in urban areas and then present a new ongoing project based on the deployment of a seismic network in educational centers located in the city of Barcelona.}
+	ABSTRACT	=	{The high degree of human activities in urban environments produces large background vibrations that makes it difficult to use data acquired in these areas for classical seismology. Seismometers installed within cities have been typically been used for the study of seismic hazard or for monitoring civil engineering problems. However, with the development of monitoring techniques based on the interpretation of the so-called seismic ambient noise, these data have gained scientific interest. Our objective is to discuss an additional utility of seismometers deployed within a city; its use as a tool to connect society with Earth sciences. Many citizen activities, from traffic to music concerts, produce vibrations that can be recorded seismically, and our experience shows that these records attract the attention of the media and social networks. With the emergence of low-cost and easy-to-use instruments in recent years, more citizens can now record ground motion and become interested in the interpretation of the recorded seismograms. The installation of permanent seismic networks in educational centers has proven to be a good approach to introduce students to Earth sciences at the national level and can also be developed at the urban scale using this new instrumentation. In this contribution we will first review the previous results related to the identification of the sources of vibration in urban areas and then present a new ongoing project based on the deployment of a seismic network in educational centers located in the city of Barcelona.}
 }
 
 @article{Subedi:2020,
@@ -115,4 +116,13 @@
 	title 		= {{ObsPy}: a bridge for seismology into the scientific Python ecosystem},
 	journal 	= {Computational Science {\&} Discovery},
[…]
