### 2021-08-11 raspishake/rsudp#22 @iannesbitt
Edits based on JOSS review
Edits based on reviewer suggestions in openjournals/joss-reviews#2565, 2021-08-10.

- [x] unit tests on a CI (already [implemented with Scrutinizer CI](https://scrutinizer-ci.com/g/raspishake/rsudp/inspections))
- [x] contributing guidelines on the github page (added [here](https://github.com/raspishake/rsudp/commit/38a95c442153b88b3e54b0db2956c95a1af809bd))
- [x] adding text from the [Why it's special](https://raspishake.github.io/rsudp/about.html#why-it-s-special) section of the docs to the paper's statement of need (added [here](https://github.com/raspishake/rsudp/commit/d5ef7b53da3621300489fbf2c6024fc59092032e))
- [x] compare to and reference other software (added [here](https://github.com/raspishake/rsudp/commit/d5ef7b53da3621300489fbf2c6024fc59092032e))
- [ ] incorporate ObsPy stream (forthcoming)
- [x] clarify research usefulness (done [here](https://github.com/raspishake/rsudp/commit/d5ef7b53da3621300489fbf2c6024fc59092032e))
- [x] add citations (added [here](https://github.com/raspishake/rsudp/commit/90f4bbdc84d85ce81e4b15817662d305840fe8d2))
