### 2021-08-11 raspishake/rsudp#22 @iannesbitt
ObsPy streaming will need to wait for a future release.
