### 2021-08-11 #2565 @iannesbitt
Hi @calum-chamberlain, thank you very much for the feedback. I have added nearly all of the suggestions you detail above, with the exception of ObsPy stream access which will have to wait for a future release. Hopefully the text I added to the Statement of Need is sufficient to detail rsudp's niche a little better. See raspishake/rsudp#22.

rsudp's research usefulness I believe is in the fact that it is lightweight and extensible and can be used to push notifications to various endpoints with very little data usage. Given limited internal storage and data restrictions, this could be useful for remote passive seismic projects that need to know when to retrieve data from their units but do not have the resources to pay for several GB/month of cellular data. The use of UDP rather than TCP is also purposeful to keep bandwidth usage down. The advantage over SeedLink is that there is no additional processing or storage read required on the part of the Shake OS, since UDP data packets are sent directly from a process in the Shake's software as data is being recorded whereas SeedLink data is provided by request. This keeps wear down slightly on the storage medium, which limits maintenance requirements for remote installations. The extensible nature of rsudp means that the STA/LTA trigger could also, for example, start a process to send event data to a server then shut it off after STA/LTA returns to normal.

If you think including this example would improve the paper, I can add it as well.
