+++ paper/paper.bib
@@ -143,7 +143,7 @@
 	abstract 	= {Swarm is a Java application designed to display and analyze seismic waveforms in real-time. Swarm can connect to and read from a variety of different static and dynamics data sources, including Earthworm & Winston wave servers, SeedLink servers, FDSN Web Services, and wave files. Swarm has both time- and frequency-domain analysis tools, along with a mapping platform. A full-screen kiosk mode allows users to monitor incoming wave and helicorder data. Swarm was originally developed by Dan Cervelli and Peter Cervelli for the USGS Volcano Science Center (VSC). More recent maintenance and enhancements were made by Tom Parker (USGS VSC AVO) and Diana Norgaard (USGS VSC VDAP), with contributions from Loren Antolik (USGS VSC HVO), Kevin Frechette (ISTI), Anthony Lomax, Dave Ketchum, Chirag Patel, and Ivan Henson.}
 }
 
-@article{Winter2021,
+@article{Winter:2021,
 	abstract 	= {We evaluate the performance of the low-cost seismic sensor Raspberry Shake (RS) to identify and monitor icequakes (which occur when glacial ice experiences brittle deformation) in extreme environments. In January 2020, three RS3D sensors were installed on a katabatic wind-scoured blue ice area (BIA) close to the Princess Elisabeth Antarctica research station in Dronning Maud Land, East Antarctica. The sensors were configured for Antarctic deployment and placed in insulated enclosures to protect them from harsh weather systems. The RS network (installed in a triangular array) performed well in the cold and with rapid air temperature change, as diurnal temperatures fluctuated from a high of 0.0°C to a minimum temperature of −15.0°C. Although battery connectivity issues in one unit limit full triangulation of seismic signals, and high background noise may mask some seismic signals, data from the RS2 unit reveals that 2936 icequakes were detected over a 10-day period. The temporal occurrence of these icequakes, combined with satellite-derived surface temperature measurements and automatic weather station data, suggest that diurnal fluctuations in solar radiation control ice surface temperature changes, driving thermal contraction of the ice. Seismic investigations like these can therefore provide information on the thermal state and ice fracture mechanics of ablation zones such as BIAs. Our work highlights the potential application of the RS (after minimal modification) in glaciated environments where equipment often needs to be portable, temporary and lightweight, and able to perform in extreme weather conditions.},
 	author 		= {Winter, Kate and Lombardi, Denis and Diaz-Moreno, Alejandro and Bainbridge, Rupert},
 	doi 		= {10.1785/0220200483},
