### 2021-08-12 #2565 @iannesbitt
Also @calum-chamberlain, I mentioned this in the linked issue but Scrutinizer CI runs the unit tests on each commit. You can see the results of this quickly by looking at the build status indicator on the main github page, or in detail by going to the [build output page](https://scrutinizer-ci.com/g/raspishake/rsudp/inspections/38728aba-a159-4082-9a19-919e9de76c30/log), clicking on "Node: tests" (on the right) then scrolling down to the output of `rs-test`
