### 2021-08-12 #2565 @iannesbitt
Tests originate [from the client](https://github.com/raspishake/rsudp/blob/master/rsudp/client.py#L489) and are defined in [test.py](https://github.com/raspishake/rsudp/blob/master/rsudp/test.py).
