### 2021-08-12 #2565 @calum-chamberlain
Thanks for that Ian - I had found that test file, but didn't think those to be unit tests - this may be a lack of understanding on my part, but the way I usually understand unit tests is a series of focused tests that test specific parts of the code base (similar to that outlined by the "agile alliance" [here](https://www.agilealliance.org/glossary/unit-test/#q=~(infinite~false~filters~(postType~(~'page~'post~'aa_book~'aa_event_session~'aa_experience_report~'aa_glossary~'aa_research_paper~'aa_video)~tags~(~'unit*20test))~searchTerm~'~sort~false~sortDirection~'asc~page~1)). Often this type of testing can be helpful to debug any issues, rather than having a fail caught somewhere in a systems test (as the main test seems to be) and needing to dig around to find the issue.  Consider this to be a suggestion for future though, rather than a requirement.

The guidelines for review do state that it it desirable for [the core functionality to be tested](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) though, and it looks like some parts of the code-base do not have tests. For example, it doesn't look like the miniseed writing module is tested at the moment, and I don't think the data forwarding, twitter or telegram modules are tested either by the current tests? Again, happy to be corrected!
