### 2021-10-01 #2565 @jedbrown
Hi, @iannesbitt. Are you intending to leave the current state of testing or will you add tests. One important aspect is to ask what components other researchers might want to contribute to and make sure they are well tested. So if you have some component that does a well-defined task and will never change, then perhaps it doesn't need its own unit test. But a leaky abstraction or one with a researchy surface really needs tests to enable community contributions.

You're the expert on this particular code, but I would like to hear your response. Generating a coverage report is one of those "that which isn't measured never gets fixed" that's likely to be informative at very low effort (https://coverage.readthedocs.io/en/coverage-5.5/).
