### 2021-10-16 d107e8e
adding write tests (openjournals/joss-reviews#2565)
