### 2021-10-16 e0c61b5
adding testing support for Twitter and Telegram modules (openjournals/joss-reviews#2565)
