### 2021-10-17 94ea93c
changing twitter/telegram tests to in-line
