### 2021-10-17 91c54f2
adding image ready message
