### 2021-10-17 9dd4f76
adding alert client tests
