+++ rsudp/c_alertsound.py
@@ -3,10 +3,11 @@ from rsudp.raspberryshake import ConsumerThread
 from rsudp import printM, printW, printE
 from rsudp.test import TEST
 import subprocess
-from tempfile import NamedTemporaryFile
 try:
 	from pydub.playback import play, PLAYER
+	from pydub import AudioSegment
 	pydub_exists = True
+	TEST['d_pydub'][1] = True
 except ImportError:
 	pydub_exists = False
 
@@ -24,7 +25,54 @@ class AlertSound(ConsumerThread):
 
 	"""
 
-	def __init__(self, testing=False, sound=False, soundloc=False, q=False):
+	def _load_sound(self):
+		'''
+		Loads MP3 sound if possible, then writes to wav.
+		Catches a ``FileNotFoundError`` when no player can be loaded.
+		'''
+		try:
+			soundloc = self.sound
+			self.sound = AudioSegment.from_file(self.sound, format="mp3")
+			printM('Loaded %.2f sec alert sound from %s' % (len(self.sound)/1000., soundloc), sender=self.sender)
+			self.wavloc = '%s.wav' % os.path.splitext(soundloc)[0]
+			if 'ffplay' in PLAYER:
+				self._write_wav()
+		except FileNotFoundError as e:
+			printE('Error loading player - %s' % (e), sender=self.sender)
+			printW("You have chosen to play a sound, but don't have ffmpeg or libav installed.", sender=self.sender)
+			printW('Sound playback requires one of these dependencies.', sender=self.sender, spaces=True)
+			printW("To install either dependency, follow the instructions at:", sender=self.sender, spaces=True)
+			printW('https://github.com/jiaaro/pydub#playback', sender=self.sender, spaces=True)
+			printW('The program will now continue without sound playback.', sender=self.sender, spaces=True)
+			self.sound = False
+
+	def _init_sound(self):
+		if pydub_exists:
+			if os.path.exists(self.sound):
+				self._load_sound()
+			else:
+				printW("The file %s could not be found." % (self.sound), sender=self.sender)
+				printW('The program will now continue without sound playback.', sender=self.sender, spaces=True)
+				self.sound = False
+		else:
+			printW("You don't have pydub installed, so no sound will play.", sender=self.sender)
+			printW('To install pydub, follow the instructions at:', sender=self.sender, spaces=True)
+			printW('https://github.com/jiaaro/pydub#installation', sender=self.sender, spaces=True)
+			printW('Sound playback also requires you to install either ffmpeg or libav.', sender=self.sender, spaces=True)
+
+	def _write_wav(self):
+		'''
+		FFPlay can only play raw wav sounds without verbosity, so to support
+		non-verbose mode we must export to .wav prior to playing a sound.
+		This function checks for an existing wav file and if it does not
+		exist, writes a new one.
+		'''
+		if not os.path.isfile(self.wavloc):
+			self.sound.export(self.wavloc, format="wav")
+			printM('Wrote wav version of sound file %s' % (self.wavloc), self.sender)
+
+
+	def __init__(self, testing=False, soundloc=False, q=False):
 		"""
 		.. _pydub.AudioSegment: https://github.com/jiaaro/pydub/blob/master/API.markdown#audiosegment
 
@@ -37,10 +85,10 @@ class AlertSound(ConsumerThread):
 		self.alive = True
 		self.testing = testing
 
-		self.sound = sound
-		self.tmpfile = None
+		self.sound = soundloc
 		self.devnull = open(os.devnull, 'w')
-		self.wavloc = '%s.wav' % os.path.splitext(soundloc)[0]
+		
[…]
