### 2021-10-17 0e0f02b
adding test to RSAM module
