### 2021-10-17 b971484
adding/fixing producer and consumer tests
