+++ rsudp/test.py
@@ -35,6 +35,7 @@ TEST = {
 	# consumers
 	'c_write':				['miniSEED write              ', False],
 	'c_miniseed':			['miniSEED data               ', False],
+	'c_print':				['print data                  ', False],
 	'c_alerton':			['alert trigger on            ', False],
 	'c_alertoff':			['alert trigger off           ', False],
 	'c_play':				['play sound                  ', False],
