### 2021-10-17 0196fb9
fixing write function and `testing` variable
