### 2021-10-17 23e1bb7
simplifying rsam printing
