+++ rsudp/c_rsam.py
@@ -22,8 +22,7 @@ class RSAM(rs.ConsumerThread):
 	``"fwport"`` with packets formatted as either JSON, "lite", or CSV.
 
 	:param queue.Queue q: queue of data and messages sent by :class:`rsudp.c_consumer.Consumer`.
-	:param bool debug: whether or not to print RSAM analysis live to the console.
-	:param bool quiet: whether or not to suppress ``debug`` printing (``True`` suppresses output).
+	:param bool quiet: ``True`` to suppress printing of RSAM analysis live to the console, ``False`` otherwise.
 	:param float interval: window of time in seconds to apply RSAM analysis.
 	:param str cha: listening channel (defaults to [S,E]HZ)
 	:param str deconv: ``'VEL'``, ``'ACC'``, ``'GRAV'``, ``'DISP'``, or ``'CHAN'``
@@ -32,7 +31,7 @@ class RSAM(rs.ConsumerThread):
 	:param str fwformat: Specify a format for the forwarded packet: ``'LITE'``, ``'JSON'``, or ``'CSV'``
 	"""
 
-	def __init__(self, q=False, debug=False, interval=5, cha='HZ', deconv=False,
+	def __init__(self, q=False, interval=5, cha='HZ', deconv=False,
 				 fwaddr=False, fwport=False, fwformat='LITE', quiet=False,
 				 testing=False,
 				 *args, **kwargs):
@@ -43,8 +42,7 @@ class RSAM(rs.ConsumerThread):
 		self.sender = 'RSAM'
 		self.alive = True
 		self.testing = testing
-		self.debug = debug
-		self.quiet = quiet	# overrides debug and suppresses printing
+		self.quiet = quiet	# suppresses printing of transmission stats
 		self.stn = rs.stn
 		self.fwaddr = fwaddr
 		self.fwport = fwport
@@ -188,7 +186,7 @@ class RSAM(rs.ConsumerThread):
 		"""
 		Print the current RSAM analysis
 		"""
-		if (self.debug) and (not self.quiet):
+		if not self.quiet:
 			msg = '%s Current RSAM: mean %s median %s min %s max %s' % (
 				(self.stream[0].stats.starttime + timedelta(seconds=
 															len(self.stream[0].data) * self.stream[0].stats.delta)).strftime('%Y-%m-%d %H:%M:%S'),
