+++ docsrc/test.rst
@@ -48,6 +48,7 @@ and their initial state:
         'c_tweetimg':           ['Twitter image message       ', False],
         'c_telegram':           ['Telegram text message       ', False],
         'c_telegramimg':        ['Telegram image              ', False],
+        'c_forward':            ['forwarding                  ', False],
         'c_rsam':               ['RSAM transmission           ', False],
         'c_custom':             ['custom code execution       ', False],
     }
