### 2021-10-17 b39e1a7
adding plot test
