+++ rsudp/test.py
@@ -33,6 +33,7 @@ TEST = {
 	'd_matplotlib':			['matplotlib backend          ', False],
 
 	# consumers
+	'c_plot':				['plot                        ', False],
 	'c_write':				['miniSEED write              ', False],
 	'c_miniseed':			['miniSEED data               ', False],
 	'c_print':				['print data                  ', False],
