### 2021-10-17 5db598c
fixing thread-unsafe variable
