### 2021-10-17 bcdd2c1
adding comment to pydub import logic
