### 2021-10-17 5cec8bd
simplifying pydub import logic
