+++ rsudp/client.py
@@ -1,5 +1,4 @@
 import sys, os
-import pathlib
 import signal
 import getopt
 import time
@@ -38,8 +37,7 @@ WRITER = False
 SOUND = False
 TESTING = False
 TESTQUEUE = False
-dp = pathlib.Path(__file__).parent.resolve()
-TESTFILE = os.path.join(dp, 'test/testdata')
+TESTFILE = pr.resource_filename('rsudp', os.path.join('test', 'testdata'))
 SENDER = 'Main'
 
 def handler(sig, frame):
