+++ MANIFEST.in
@@ -1,3 +1,3 @@
 include rsudp/rs_sounds/*.mp3
 include rsudp/img/*
-include rsudp/test/testdata
+include rsudp/test/testdata.ms
