### 2021-10-18 #2565 @iannesbitt
Hi all, the requested changes have been added and merged to master. In particular:
- all core modules are tested explicitly now, if applicable
- code coverage is now implemented using python-cc (a badge has been added to represent this)

Apologies for the delay...I needed ample time to make sure this was done as thoroughly as possible.
If I missed anything, let me know. Cheers and thank you again for your attention to detail.
