+++ setup.py
@@ -17,7 +17,7 @@ setuptools.setup(
     url="https://github.com/raspishake/rsudp",
     packages=setuptools.find_packages(),
     include_package_data=True,
-    install_requires=['obspy', 'numpy', 'matplotlib==3.1.1', 'pydub', 'twython',
+    install_requires=['obspy', 'numpy', 'matplotlib<3.2', 'pydub', 'twython',
                       'python-telegram-bot'],
     entry_points = {
         'console_scripts': [
