+++ CHANGES.md
@@ -1,6 +1,7 @@
 # Changelog
 ## changes in 1.0.3
 - changed matplotlib pin to be `<3.2` rather than `==3.1.1` to address [#21](https://github.com/raspishake/rsudp/issues/21)
+- modified logos
 
 ## changes in 1.0.2
 - corrected install script to fix [#14](https://github.com/raspishake/rsudp/issues/14)
