### 2021-10-29 734ff9c
adding version to logging sequence
