### 2021-10-29 7238642
addressing #27 (unicode output error)
