+++ CHANGES.md
@@ -2,6 +2,8 @@
 ## changes in 1.0.3
 - changed matplotlib pin to be `<3.2` rather than `==3.1.1` to address [#21](https://github.com/raspishake/rsudp/issues/21)
 - modified logos
+- fixed unicode output error (emoji caused error on Windows machines)
+- added version printout to rs lib initialization sequence
 
 ## changes in 1.0.2
 - corrected install script to fix [#14](https://github.com/raspishake/rsudp/issues/14)
