### 2021-10-29 b90cedf
simplifying imports in consumer modules
