+++ rsudp/packetloss.py
@@ -1,6 +1,5 @@
 import sys
 import getopt
-import datetime as dt
 import signal
 from rsudp import raspberryshake, printM, printW, printE, add_debug_handler
 
