+++ rsudp/raspberryshake.py
@@ -1,6 +1,3 @@
-from datetime import datetime, timedelta
-import time
-import math
 import numpy as np
 import sys, os, platform
 import socket as s
