+++ rsudp/c_telegram.py
@@ -1,7 +1,5 @@
 import os, sys
 import time
-from datetime import datetime, timedelta
-from rsudp.raspberryshake import ConsumerThread
 import rsudp.raspberryshake as rs
 from rsudp import printM, printW, printE, helpers
 from rsudp.test import TEST
