### 2021-10-29 a5e1e1b
adding extra text feature to telegram module (#28)
