### 2021-10-29 eaebecf
simplifying init sequences
