+++ rsudp/c_tweet.py
@@ -69,6 +69,7 @@ class Tweeter(rs.ConsumerThread):
 		Initialize the process
 		"""
 		super().__init__()
+		self.queue = q
 		self.sender = 'Tweeter'
 		self.alive = True
 		self.tweet_images = tweet_images
@@ -83,25 +84,7 @@ class Tweeter(rs.ConsumerThread):
 
 		self._resolve_extra_text(extra_text)
 
-		if q:
-			self.queue = q
-		else:
-			printE('no queue passed to consumer! Thread will exit now!', self.sender)
-			sys.stdout.flush()
-			self.alive = False
-			sys.exit()
-
-		if not self.testing:
-			self.twitter = Twython(
-				consumer_key,
-				consumer_secret,
-				access_token,
-				access_token_secret
-			)
-		else:
-			printW('The Twitter module will not post to Twitter in Testing mode.',
-					self.sender, announce=False)
-
+		self.auth()
 
 		self.livelink = u'live feed ➡️ https://stationview.raspberryshake.org/#?net=%s&sta=%s' % (rs.net, rs.stn)
 		self.message0 = '(#RaspberryShake station %s.%s%s) Event detected at' % (rs.net, rs.stn, self.region)
@@ -126,12 +109,16 @@ class Tweeter(rs.ConsumerThread):
 
 
 	def auth(self):
-		self.twitter = Twython(
-			self.consumer_key,
-			self.consumer_secret,
-			self.access_token,
-			self.access_token_secret
-		)
+		if not self.testing:
+			self.twitter = Twython(
+				self.consumer_key,
+				self.consumer_secret,
+				self.access_token,
+				self.access_token_secret
+			)
+		else:
+			printW('The Twitter module will not post to Twitter in Testing mode.',
+					self.sender, announce=False)
 
 
 	def getq(self):
