### 2021-10-29 66e57b4
moving duplicate function to helpers
