### 2021-10-29 843039e
adding versionadded to extra text helper function docstring
