### 2021-10-29 52ab9dd
unhiding extra text helper function
