### 2021-10-29 f2813a5
adding returns to extra text helper function docstring
