### 2021-11-01 6c68ab8
fixing the packetization script
