### 2021-11-01 83ec6e6
fixing coverage of packetize script
