+++ rsudp/client.py
@@ -510,6 +510,7 @@ default settings and the data file at
 	settings_are_default = True
 	plot = True
 	quiet = False
+	customfile = False
 
 	try:
 		opts = getopt.getopt(sys.argv[1:], 'hf:s:bq',
@@ -535,13 +536,11 @@ default settings and the data file at
 					out = '%s.txt' % (a)
 					packetize(inf=a, outf=out, testing=True)
 					TESTFILE = out
+					customfile = True # using a custom miniseed file for testing
 				except Exception as e:
 					print(hlp_txt)
 					print(COLOR['red'] + 'ERROR: %s' % e + COLOR['white'])
 					exit(1)
-		else:
-			# else we are just using the default miniseed file
-			packetize(inf=TESTFILE+'.ms', outf=TESTFILE, testing=True)
 		if o in ('-s', '--settings='):
 			'''
 			Dump the settings to a file, specified after the `-d` flag, or `-d default` to let the software decide where to put it.
@@ -558,6 +557,9 @@ default settings and the data file at
 		if o in ('-q', '--no-sound'):
 			quiet = True
 
+	if not customfile:
+		# we are just using the default miniseed file
+		packetize(inf=TESTFILE+'.ms', outf=TESTFILE, testing=True)
 
 	T.TEST['n_internet'][1] = T.is_connected('www.google.com')
 
