### 2021-11-02 #2565 @calum-chamberlain
Hi all, apologies, but I'm currently in the field installing seismometers along the Alpine Fault and won't be able to get to this until December. Sorry!
