### 2021-11-03 3396a1b
adjusting default settings for easier quickstart
