+++ rsudp/helpers.py
@@ -40,12 +40,12 @@ def default_settings(output_dir='%s/rsudp' % os.path.expanduser('~').replace('\\
     "channels": ["all"]},
 "plot": {
     "enabled": true,
-    "duration": 30,
+    "duration": 90,
     "spectrogram": true,
-    "fullscreen": false,
+    "fullscreen": true,
     "kiosk": false,
     "eq_screenshots": false,
-    "channels": ["HZ", "HDF"],
+    "channels": ["all"],
     "deconvolve": true,
     "units": "CHAN"},
 "forward": {
@@ -60,10 +60,10 @@ def default_settings(output_dir='%s/rsudp' % os.path.expanduser('~').replace('\\
     "channel": "HZ",
     "sta": 6,
     "lta": 30,
-    "threshold": 1.7,
-    "reset": 1.6,
-    "highpass": 0,
-    "lowpass": 50,
+    "threshold": 3.95,
+    "reset": 0.9,
+    "highpass": 0.8,
+    "lowpass": 9,
     "deconvolve": false,
     "units": "VEL"},
 "alertsound": {
