+++ docsrc/settings.rst
@@ -634,12 +634,12 @@ By default, the settings are as follows:
         "channels": ["all"]},
     "plot": {
         "enabled": true,
-        "duration": 30,
+        "duration": 90,
         "spectrogram": true,
-        "fullscreen": false,
+        "fullscreen": true,
         "kiosk": false,
         "eq_screenshots": false,
-        "channels": ["HZ", "HDF"],
+        "channels": ["all"],
         "deconvolve": true,
         "units": "CHAN"},
     "forward": {
@@ -654,10 +654,10 @@ By default, the settings are as follows:
         "channel": "HZ",
         "sta": 6,
         "lta": 30,
-        "threshold": 1.7,
-        "reset": 1.6,
-        "highpass": 0,
-        "lowpass": 50,
+        "threshold": 3.95,
+        "reset": 0.9,
+        "highpass": 0.8,
+        "lowpass": 9,
         "deconvolve": false,
         "units": "VEL"},
     "alertsound": {
