### 2021-11-03 0fe6ab7
adjusting default settings for easier quickstart
