### 2021-11-04 #2565 @jedbrown
Thanks, @iannesbitt. I think this is a great improvement. I know this has been in review for a long time so I'd like to offer @calum-chamberlain a choice:
1. we'll wait for you to check in in December (default if you can't reply)
2. you can defer to me on this point and I'll check that your recommendations have been reasonably met, and if so, proceed with any final copy edits and acceptance. (Note that your checklist has unchecked "Statement of need" and "State of the field".)
