### 2021-11-04 cc7f56f
adding convenience scripts, changing default alert settings
