+++ setup.py
@@ -25,6 +25,9 @@ setuptools.setup(
             'rs-client=rsudp.client:main',
             'rs-test=rsudp.client:test',
             'packetize=rsudp.packetize:main',
+            'rs-settings=rsudp.helpers:ep_edit_settings',
+            'rs-log=rsudp.helpers:ep_cat_log',
+            'rs-tailf=rsudp.helpers:ep_tailf_log',
             ],
     },
     classifiers=[
