+++ docs/_sources/index.rst.txt
@@ -74,7 +74,7 @@ The modules available in rsudp are organized by type below.
     init
     raspberryshake
     helpers
-    entrypoints
+    entry_points
 
 .. toctree::
     :maxdepth: 2
