### 2021-11-04 3c18798
removing errant printE call
