### 2021-11-04 b67864d
doc corrections
