### 2021-11-04 c20ca6e
doc corrections
