### 2021-12-05 #2565 @calum-chamberlain
Hi all, sorry for the late reply - @jedbrown this looks great to me. @jedbrown I'm so sorry this took so long - the tests and the coverage look great, thank you for adding that, I really hope it does help in the long-run to have your code tested.

@jedbrown I approve this revision - I think that I have ticked everything that I need to, and I have reviewed the most recent version of the manuscript. Do you need anything else from me?
