### 2021-12-09 #2565 @iannesbitt
Hi @jedbrown, here is the tag and DOI record.

Release: [1.1.0](https://github.com/raspishake/rsudp/releases/tag/1.1.0)
Zenodo: https://zenodo.org/record/5771026
doi: [10.5281/zenodo.5771026](https://doi.org/10.5281/zenodo.5771026)
