### 2021-12-13 #2565 @iannesbitt
@jedbrown a reader pointed out a small typographical error in the references: "Geolgical" should be "Geological". Wondering if it's too late to change now. I've made the necessary edit to the .bib file.
